+++ crates/aiken/src/cmd/blueprint/apply.rs
@@ -48,7 +48,7 @@ pub fn exec(
         validator,
     }: Args,
 ) -> miette::Result<()> {
-    with_project(None, |p| {
+    with_project(None, false, |p| {
         let title = module.as_ref().map(|m| {
             format!(
                 "{m}{}",